+++ redisson-spring-boot-autoconfigure/src/main/java/com/github/trang/redisson/autoconfigure/enums/RedissonType.java
@@ -7,15 +7,29 @@ package com.github.trang.redisson.autoconfigure.enums;
  */
 public enum RedissonType {
 
-    /** 单节点模式 */
+    /**
+     * 单节点模式
+     */
     SINGLE,
-    /** 集群模式 */
+
+    /**
+     * 集群模式
+     */
     CLUSTER,
-    /** 主从模式 */
+
+    /**
+     * 主从模式
+     */
     MASTER_SLAVE,
-    /** 哨兵模式 */
+
+    /**
+     * 哨兵模式
+     */
     SENTINEL,
-    /** 云托管模式 */
+
+    /**
+     * 云托管模式
+     */
     REPLICATED
 
 }